refactor: improve login component
