+++ src/components/Login/index.ts
@@ -1,1 +1,1 @@
-export * from './LoginPortal'
+export * from './Login'
